backend/requirements.txt
django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
django-autoslug
pyjwt
django-countries
pillow
django-filter